I made this PR to fix #197.

I use promote_enums  in QtTest.
In Current QtPy, promote_enums is used in QtCore/QtGui/QtWidgets only, but is there any reason?